fix:‌fix design issues and refactor
