+++ components/organisms/cardCarousel/components/cardCarouselController.tsx
@@ -1,8 +1,4 @@
-import Heading from "@/components/atoms/heading/heading";
 import SeeAll from "@/components/atoms/link/link";
-import Button from "@/components/atoms/button/button";
-import Row from "@/components/atoms/grid/row";
-import Column from "@/components/atoms/grid/column";
 import { CardControllerWrapper } from "./cardCarousel.styles";
 import Buttons from "@/components/molecules/buttonGroup/buttonGroup";
