+++ components/organisms/slideshow/sliderShow.style.ts
@@ -1,6 +1,6 @@
 import styled, { css } from "styled-components";
-import Button from "../atoms/button/button";
-import Dots from "./cardCarousel/components/dots";
+import Button from "../../atoms/button/button";
+import Dots from "../cardCarousel/components/dots";
 
 export const MainWrapper = styled.div`
   position: relative;
@@ -21,27 +21,25 @@ export const SliderWrapper = styled.div`
   -ms-overflow-style: none;
   scrollbar-width: none;
 
-  ${({ theme }) => css`
-    position: relative;
-    width: 100%;
-    height: 400px;
-    background-color: ${({ theme }) => theme.colors.yellow};
+  position: relative;
+  width: 100%;
+  height: 400px;
+  background-color: ${({ theme }) => theme.colors.warning};
 
-    @media (min-width: ${({ theme }) => theme.breakpoints.sm}) {
-      height: 440px;
-    }
+  @media (min-width: ${({ theme }) => theme.breakpoints.sm}) {
+    height: 440px;
+  }
 
-    @media (min-width: ${({ theme }) => theme.breakpoints.md}) {
-      height: 560px;
-    }
+  @media (min-width: ${({ theme }) => theme.breakpoints.md}) {
+    height: 560px;
+  }
 
-    @media (min-width: ${({ theme }) => theme.breakpoints.lg}) {
-      height: 600px;
-    }
-    @media (min-width: ${({ theme }) => theme.breakpoints.xl}) {
-      height: 696px;
-    }
-  `}
+  @media (min-width: ${({ theme }) => theme.breakpoints.lg}) {
+    height: 600px;
+  }
+  @media (min-width: ${({ theme }) => theme.breakpoints.xl}) {
+    height: 696px;
+  }
 `;
 export const Controller = styled.div`
   position: absolute;
